requirements/base.txt
Django==5.0.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
drf-spectacular==0.27.2
django-environ==0.11.2
Pillow==10.4.0
psycopg2-binary==2.9.11
celery==5.4.0
redis==5.1.1
django-celery-results==2.5.1
django-rest-passwordreset==1.4.2
PyYAML==6.0.2
requests==2.32.3
django-cors-headers==4.3.1
